Use DEB repo instead of putting binaries in git, cleanups
Storing the full tyk distribution on git is onerous and increases the image complexity so replaced that with using a DEB repo as we do with other public docker images.

Changing base image to debian:jessie-slim as it's twice smaller than ubuntu:14.04, while having a lot less vulnerabilities according to Docker Hub. This is about the same level of packages version as Ubuntu 14.04 too. In fact, it has more recent patch-versions of redis and nginx within the same minor.

Replacing manual protobuf build with downloading binary "wheels" brings the largest benefits in terms of image size, complexity and time to build.

Added a .dockerignore file to significantly decrease the size of the build context.

Having all this along with various cleanups and improvements this dramatically decreases the image size and complexity, makes release management easier too.